[jetty9] use jetty9.3 as embedded jetty
+ use jetty9.3
+ adapt embedded library strategy
++ support websockets
+ cleanup maven dependencies in jetty embedded
