+++ plugins/eclipse-jetty-launcher/src/main/resources/net/sourceforge/eclipsejetty/messages.properties
@@ -50,7 +50,7 @@ advConfigTab_customWebDefaultsWorkspaceButton=Workspace...
 advConfigTab_customWebDefaultsWorkspaceButtonTip=Search the workspace for a custom web defaults XML file.
 advConfigTab_customWebDefaultsWorkspaceMessage=Select a resource as Jetty Context file:
 advConfigTab_customWebDefaultsWorkspaceTitle=Resource Selection
-advConfigTab_embeddedButton=Use Jetty provided by launcher plugin (Jetty 8.x)
+advConfigTab_embeddedButton=Use Jetty provided by launcher plugin (Jetty 9.3)
 advConfigTab_embeddedButtonTip=The Eclipse Jetty Plugin comes with a tested default Jetty. Use it\!
 advConfigTab_externalButton=External...
 advConfigTab_externalButtonTip=Search for a Jetty installation in your file system.
